+++ fieldmask/src/lib.rs
@@ -137,4 +137,6 @@ mod maskable;
 pub use fieldmask_derive::{Maskable, OptionMaskable, SelfMaskable, maskable_atomic};
 
 pub use mask::{Mask, MaskInput};
-pub use maskable::{DeserializeMaskError, Maskable, OptionMaskable, SelfMaskable, UpdateOptions};
+pub use maskable::{
+    DeserializeMaskError, Maskable, OptionMaskable, ProjectOptions, SelfMaskable, UpdateOptions,
+};